build: utils cmodels file and create id cmodels method in usecases
