+++ src/use_cases/cmodels.py
@@ -7,3 +7,7 @@ class CmodelUseCases:
         model_in_db.pop('inserted_at')
         model_in_db.pop('updated_at')
         return model_in_db
+
+    def create_id_cmodel(model_name: str) -> dict:
+
+        return {'_id': model_name.encode('utf-8').hex()}
